+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: mstupnik <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/18 13:07:30 by mstupnik          #+#    #+#             */
-/*   Updated: 2019/09/19 13:39:29 by mstupnik         ###   ########.fr       */
+/*   Updated: 2019/09/19 14:36:40 by mstupnik         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	*ft_memmove(void *dst, const void *src, size_t len)
 	char	*cdst;
 	char	*csrc;
 	char	temp[len];
-	int		i;
+	size_t	i;
 
 	cdst = (char *)dst;
 	csrc = (char *)src;
